Bclerc

+++ back/src/wschat/wschat.service.ts
@@ -12,6 +12,8 @@ import { BasicUserI } from 'src/user/interface/basicUser.interface';
 import { UserService } from 'src/user/user.service';
 import { EventEmitter2 } from '@nestjs/event-emitter';
 import { PenaltiesService } from 'src/chat/services/penalties/penalties.service';
+import e from 'express';
+import { PusnishI } from 'src/chat/interfaces/punish.interface';
 
 @Injectable()
 export class WschatService {
@@ -39,15 +41,14 @@ export class WschatService {
   async subscribeToRoom(socketId: string, subRoom: SubscribeRoomDto) {
     const user = this.onlineUserService.getUser(socketId);
     const room = await this.chatService.getRoomById(subRoom.roomId);
-    const penalty = await this.penaltiesService.getRoomPenaltiesForUser(user.id, room.id);
-    
-    
+    const penalty = await this.penaltiesService.getRoomPenaltiesForUser(user.id, room.id);    
+    console.log(penalty);
     if (penalty && penalty.type === PenaltyType.BAN) {
       this.eventEmitter.emit('room.user.join', {
         room: room,
         user: user,
         success: false,
-        message: 'Vous avez été banni de ce salon' + (penalty.timetype === 'PERM' ? '.' : ' jusqu\'au ' + penalty.endTime)
+        message: 'Vous avez été banni de ce salon' + (penalty.timetype === 'PERM' ? '.' : ' jusqu\'au ' + penalty.endTime.toLocaleString('fr-FR'))
       });
       return ;
     }
@@ -72,12 +73,37 @@ export class WschatService {
     }
   }
 
+  async punishUser(socketId: string, event: PusnishI) {
+    const user = this.onlineUserService.getUser(socketId);
+    const target = await this.userService.findOne(event.targetId);
+    const room = await this.chatService.getRoomById(event.roomId);
+    const now = new Date().getTime();
+    let endTime: Date;
+
+    if (user && target && room) {
+      if (room.admins.find(admin => admin.id == user.id)) {
+        if (room.ownerId != target.id) {
+          endTime = new Date(now + event.time);
+          this.penaltiesService.punishUser(target.id, room.id, event.type, event.perm ? null : endTime);
+          this.eventEmitter.emit('room.user.punished', 
+          { room: room, user: target, punisher: user, type: event.type, success: true});
+        } else {
+          this.eventEmitter.emit('room.user.punished',
+            { room: room, user: target, punisher: user, type: event.type, success: false, message: 'Vous ne pouvez pas punir le propriétaire du salon'});
+        }
+      } else {
+        this.eventEmitter.emit('room.user.punished', 
+          { room: room, user: target, punisher: user, type: event.type, success: false, message: 'Vous n\'êtes pas administrateur de ce salon'});
+      }
+    }
+  }
+
   async ejectUserFromRoom(socketId: string, room: EjectRoomI) {
     const user = this.onlineUserService.getUser(socketId);
     const target = await this.userService.findOne(room.targetId);
     const roomToEject = await this.chatService.getRoomById(room.roomId);
 
-    if (user) {
+    if (user && target.id != roomToEject.ownerId) {
       if (roomToEject.admins.find(admin => admin.id == user.id)) {
         await this.chatService.removeUsersFromRoom(roomToEject.id, target.id);
         this.eventEmitter.emit('room.user.kicked', { room: roomToEject, user: target, kicker: user });
@@ -110,7 +136,7 @@ export class WschatService {
     const roomToDemote = await this.chatService.getRoomById(event.roomId);
 
     if (user) {
-      if (roomToDemote.ownerId == user.id) {
+      if (roomToDemote.ownerId == user.id && roomToDemote.ownerId != target.id) {
         this.chatService.removeAdminsFromRoom(roomToDemote.id, target.id);
         this.updateRoomForUsersInRoom(roomToDemote.id);
         this.eventEmitter.emit('room.admin.update', {
@@ -149,34 +175,41 @@ export class WschatService {
     const room = await this.chatService.getRoomById(roomId);
     
     if (user) {
-      await this.chatService.removeUsersFromRoom(room.id, user.id);
+      const newRoom = await this.chatService.removeUsersFromRoom(room.id, user.id);
+      if (room.ownerId == user.id) {
+        if (newRoom.users.length > 0) {
+          this.chatService.updateRoomOwner(room.id, newRoom.users[0].id);
+          this.updateRoomForUsersInRoom(room.id);
+        } else {
+          // delete
+          this.eventEmitter.emit('room.delete', { room: room });
+        }
+      }
       this.eventEmitter.emit('room.user.leave', { room: room, user: user });
     }
   }
 
-
   async editRoom(socketId: string, newRoom: newChatRoomI){
     const user = this.onlineUserService.getUser(socketId);
     const room = await this.chatService.getRoomById(newRoom.id);
+    if (!newRoom.name
+        || newRoom.name.length < 3
+        || newRoom.name.length > 20
+        || !newRoom.description
+        || newRoom.description.length < 3
+        || newRoom.description.length > 100)
+      {
+        this.eventEmitter.emit('room.update', {user: user, room: room, success: false, message: 'Les champs ne sont pas valides' });
+        return ;
+      }
     if (user) {
       if (room.ownerId == user.id) {
         await this.chatService.editRoom(newRoom);
-        this.eventEmitter.emit('room.update', { room: newRoom });
+        this.eventEmitter.emit('room.update', {user: user, room: newRoom, success: true });
+      } else {
+        this.eventEmitter.emit('room.update', {user: user, room: room, success: false, message: 'Seul le propriétaire du salon peut le modifier' });
       }
     }
-  }
-
-  async updateUserRooms(user: BasicUserI) {
-
-    const rooms = await this.chatService.getRoomsFromUser(user.id);
-    this.sendToUser(user, 'rooms', rooms);
-  }
-
-  async updateUsersMessagesInRoom (room: ChatRoomI) {
-      for (let user of room.users) {
-        const messages = await this.chatService.getMessagesFromRoomId(user.id, room.id);
-        this.sendToUser(user, 'messages', messages);
-      }
   }
 
   async updateRoomForUsersInRoom(roomId: number) {
@@ -189,15 +222,7 @@ export class WschatService {
   }
 
   sendToUser(user: BasicUserI, prefix: string, data: any) {
-    this.onlineUserService.sendToUser(user, prefix, data);
-  }
-
-  async sendToUsersInRoom(roomId: number, prefix: string, data: any) {
-    let room = await this.chatService.getRoomById(roomId);
-
-    room.users.forEach(user => {
-      this.sendToUser(user, prefix, data);
-    });
+    this.sendToUser(user, prefix, data);
   }
